+++ lib/include/pl/helpers/types.hpp
@@ -9,7 +9,7 @@ namespace pl {
     using u16  = std::uint16_t;
     using u32  = std::uint32_t;
     using u64  = std::uint64_t;
-    using u128 = __uint128_t;
+    using u64 = __uint128_t;
 
     using i8   = std::int8_t;
     using i16  = std::int16_t;
