+++ lib/include/pl/patterns/pattern_signed.hpp
@@ -35,7 +35,7 @@ namespace pl::ptrn {
             auto data = this->getValue().toSigned();
             auto size = this->getSize();
 
-            return Pattern::formatDisplayValue(fmt::format("{:d} (0x{:0{}X})", data, u128(data) & hlp::bitmask(8 * size), size * 2), this->getValue());
+            return Pattern::formatDisplayValue(fmt::format("{:d} (0x{:0{}X})", data, u64(data) & hlp::bitmask(8 * size), size * 2), this->getValue());
         }
 
         [[nodiscard]] std::string toString() const override {
